+++ tests/normalization.rs
@@ -140,6 +140,29 @@ mod minmax_tests {
     }
 }
 
+mod ocra_tests {
+    use super::*;
+
+    #[test]
+    fn test_normalize() -> Result<(), McdmError> {
+        let matrix = dmatrix![
+            2.9, 2.31, 0.56, 1.89;
+            1.2, 1.34, 0.21, 2.48;
+            0.3, 2.48, 1.75, 1.69
+        ];
+        let criteria_types = CriteriaType::from(vec![-1, 1, 1, -1])?;
+        let normalized_matrix = matrix.normalize_ocra(&criteria_types)?;
+        let expected_matrix = dmatrix![
+            0.0, 0.7238806, 1.66666667, 0.34911243;
+            5.66666667, 0.0, 0.0, 0.0;
+            8.66666667, 0.85074627, 7.33333333, 0.46745562
+        ];
+        assert_relative_eq!(normalized_matrix, expected_matrix, epsilon = 1e-5);
+
+        Ok(())
+    }
+}
+
 mod sum_tests {
     use super::*;