+++ tests/ranking.rs
@@ -358,6 +358,29 @@ mod moora_tests {
     }
 }
 
+mod ocra_tests {
+    use super::*;
+
+    #[test]
+    fn test_rank() -> Result<(), McdmError> {
+        let matrix = dmatrix![
+            2.9, 2.31, 0.56, 1.89;
+            1.2, 1.34, 0.21, 2.48;
+            0.3, 2.48, 1.75, 1.69
+        ];
+        let weights = dvector![0.25, 0.25, 0.25, 0.25];
+        let criteria_types = mcdm::CriteriaType::from(vec![-1, 1, 1, -1])?;
+        let ranking = matrix.rank_ocra(&criteria_types, &weights)?;
+        assert_relative_eq!(
+            ranking,
+            dvector![0.0, 0.73175174, 3.64463555],
+            epsilon = 1e-5
+        );
+
+        Ok(())
+    }
+}
+
 mod topsis_tests {
     use super::*;