+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Ranking methods: ERVD, MAIRCA, MARCOS, MOORA
+- Ranking methods: ERVD, MAIRCA, MARCOS, MOORA, OCRA
+- OCRA-specific normalization method
 
 ## [0.2.0] - 2024-11-30